Enhance deleteLending method in LendingController with validation for lendingId and error handling for exceptions

+++ src/main/java/lk/puLeeNa/LibManagement/controller/LengdingController.java
@@ -39,8 +39,19 @@ public class LengdingController {
     }
     @DeleteMapping
     public ResponseEntity<Void> deleteLending(@RequestParam ("lendingId") String lendingId){
-        lendingService.deleteLendingData(lendingId);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        if(lendingId == null){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        try{
+            lendingService.deleteLendingData(lendingId);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }catch (LendingDataNotFoundException e){
+            e.printStackTrace();
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }catch (Exception e){
+            e.printStackTrace();
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
     @PatchMapping
     public ResponseEntity<Void> handOverBook(@RequestParam ("lendingId") String lendingId){
